RENAME to addons/contexts && REMOVE files no longer needed in the monorepo && REMOVE unneeded dependencies & scripts
