--- addons/addon-contexts/.babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/env",
-      {
-        "modules": false
-      }
-    ],
-    "@babel/typescript",
-    "@babel/react"
-  ],
-  "comments": false
-}
