--- addons/addon-contexts/.prettierrc
@@ -1,13 +1,0 @@
-{
-  "printWidth": 100,
-  "tabWidth": 2,
-  "useTabs": false,
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "es5",
-  "bracketSpacing": true,
-  "jsxBracketSameLine": false,
-  "arrowParens": "always",
-  "proseWrap": "never",
-  "endOfLine": "lf"
-}
