--- addons/addon-contexts/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "addon-contexts",
-  "version": "1.2.1",
-  "description": "Storybook Addon Contexts",
-  "keywords": [
-    "storybook",
-    "react",
-    "vue"
-  ],
-  "author": "Leo Y. Li",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/**/*",
-    "register.js",
-    "react.js",
-    "vue.js"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/leoyli/addon-contexts.git"
-  },
-  "scripts": {
-    "prettier": "prettier --config .prettierrc --write '**/*.{ts,tsx}'",
-    "type:check": "tsc --isolatedModules --noEmit ",
-    "type:watch": "yarn type:check --watch",
-    "type:emit": "rm -rf dist && tsc --declaration --removeComments --emitDeclarationOnly",
-    "test": "jest --passWithNoTests",
-    "build": "babel src --out-dir dist --extensions \".ts,.tsx\" --ignore src/@types",
-    "prepare": "yarn test && yarn type:emit && yarn build"
-  },
-  "dependencies": {
-    "@storybook/addons": "^5.0.0",
-    "@storybook/components": "^5.0.0",
-    "@storybook/core-events": "^5.0.0"
-  },
-  "devDependencies": {
-    "@babel/cli": "7.4.3",
-    "@babel/core": "7.4.3",
-    "@babel/preset-env": "7.4.3",
-    "@babel/preset-react": "7.0.0",
-    "@babel/preset-typescript": "7.3.3",
-    "@types/jest": "24.0.11",
-    "@types/node": "^11.13.2",
-    "@types/react": "16.8.12",
-    "@types/vue": "^2.0.0",
-    "jest": "24.7.1",
-    "jest-dom": "3.1.3",
-    "jest-haste-map": "24.7.1",
-    "jest-resolve": "24.7.1",
-    "prettier": "1.16.4",
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-testing-library": "6.1.2",
-    "typescript": "3.4.2",
-    "vue": "^2.6.10"
-  },
-  "peerDependencies": {
-    "@storybook/addons": ">=5 <6",
-    "@storybook/components": ">=5 <6"
-  }
-}
