--- addons/addon-contexts/react.js
@@ -1,3 +1,0 @@
-import { withContexts } from './dist/preview/frameworks/react';
-export { withContexts };
-export default withContexts;
