--- addons/addon-contexts/register.js
@@ -1,1 +1,0 @@
-export * from './dist/register';
