--- addons/addon-contexts/src/@types/_storybook__components.d.ts
@@ -1,1 +1,0 @@
-declare module '@storybook/components';
