--- addons/addon-contexts/src/index.ts
@@ -1,8 +1,0 @@
-import { withContexts } from './preview/frameworks/react';
-export { withContexts };
-export default withContexts;
-
-console.error(
-  "[addon-contexts] Deprecation warning: `import { withContexts } from 'addon-contexts'` has been " +
-    "deprecated. Please import from 'addon-contexts/react' instead."
-);
