--- addons/addon-contexts/src/preview/libs/index.ts
@@ -1,4 +1,0 @@
-export { memorize, singleton } from './functionals';
-export { getContextNodes } from './getContextNodes';
-export { getPropsMap } from './getPropsMap';
-export { getRendererFrom } from './getRendererFrom';
