--- addons/addon-contexts/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "dist",
-    "module": "commonjs",
-    "jsx": "react",
-    "target": "esnext",
-    "moduleResolution": "node",
-    "strict": true,
-    "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true
-  },
-  "include": ["src"],
-  "exclude": ["src/register.ts"]
-}
