--- addons/addon-contexts/vue.js
@@ -1,3 +1,0 @@
-import { withContexts } from './dist/preview/frameworks/vue';
-export { withContexts };
-export default withContexts;
