+++ addons/contexts/src/manager/ToolbarControl.tsx
@@ -18,7 +18,7 @@ export const ToolbarControl: TToolbarControl = ({
     // validate the selected name
     (paramNames.concat(OPT_OUT).includes(selected) && selected) ||
     // fallback to default
-    (params.find(param => !!param.default) || { name: null }).name ||
+    (params.find((param) => !!param.default) || { name: null }).name ||
     // fallback to the first
     params[0].name;
   const list = options.cancelable === false ? paramNames : [OPT_OUT, ...paramNames];
