+++ addons/contexts/src/preview/frameworks/vue.ts
@@ -8,9 +8,9 @@ import { Renderer } from '../../@types';
  * This is the framework specific bindings for Vue.
  * '@storybook/vue' expects the returning object from a decorator to be a 'VueComponent'.
  */
-export const renderVue: Renderer = (nodes, propsMap, next) => {
+export const renderVue: Renderer = (nodes, props, next) => {
   const { getRendererFrom, updateReactiveSystem } = addonContextsAPI();
-  const reactiveProps = updateReactiveSystem(propsMap);
+  const reactiveProps = updateReactiveSystem(props);
   return Vue.extend({
     name: ID,
     data: () => reactiveProps,
